+++ janusgraph-core/src/main/java/org/janusgraph/graphdb/query/condition/PredicateCondition.java
@@ -38,7 +38,9 @@ public class PredicateCondition<K, E extends JanusGraphElement> extends Literal<
     private final Object value;
 
     public PredicateCondition(K key, JanusGraphPredicate predicate, Object value) {
-        Preconditions.checkArgument(key instanceof String || key instanceof RelationType);
+        if (key != null) {
+            Preconditions.checkArgument(key instanceof String || key instanceof RelationType);
+        }
         this.key = key;
         this.predicate = Preconditions.checkNotNull(predicate);
         this.value = value;
@@ -51,6 +53,10 @@ public class PredicateCondition<K, E extends JanusGraphElement> extends Literal<
 
     @Override
     public boolean evaluate(E element) {
+        if (key == null) {
+            return false;
+        }
+
         RelationType type;
         if (key instanceof String) {
             type = ((InternalElement) element).tx().getRelationType((String) key);